Fix Welcomer and change the database
